Modularize w/ slighlty better error handling
