+++ native/src/meetup/stellar/mod.rs
@@ -1,0 +1,1 @@
+pub mod cryptocompare;
